+++ program/source/scene.cpp
@@ -118,14 +118,14 @@ namespace cse
     hooks.call<void(const bool *)>(hook::POST_INPUT, input);
   }
 
-  void scene::simulate(const float poll_rate)
+  void scene::simulate(const double poll_rate)
   {
     if (state.active.phase != help::phase::CREATED) throw exception("Scene must be created before simulation");
-    hooks.call<void(const float)>(hook::PRE_SIMULATE, poll_rate);
+    hooks.call<void(const double)>(hook::PRE_SIMULATE, poll_rate);
     timers.update(poll_rate);
     state.active.camera->simulate(poll_rate);
     for (const auto &[name, object] : state.active.objects) object->simulate(poll_rate);
-    hooks.call<void(const float)>(hook::POST_SIMULATE, poll_rate);
+    hooks.call<void(const double)>(hook::POST_SIMULATE, poll_rate);
   }
 
   void scene::render(SDL_GPUDevice *gpu, SDL_GPUCommandBuffer *command_buffer, SDL_GPURenderPass *render_pass,
